Add FAQ entry about dedicated billing accounts
A the moment, FlexSave is only available when a customer's AWS account
is consolidated with the DoiT International billing organization.

This is mentioned in the intro. However, some customers with dedicated
billing accounts may miss that detail and wonder why they do not have
the option to enable FlexSave.

To address that problem, this change adds an FAQ entry with instructions
to sign up for an email notification when dedicated billing accounts are
supported.

In addition, I modified the overview page for consistency and style:

- I changed "account/s" to "account". The former is considered poor
  style (see https://developers.google.com/style/plurals-parentheses).
  In this instance, we can use the singular and the plural is implied.

- Mirroring the FAQ, I added a link to the AWS documentation
  consolidated billing

This fixes CMP-1605