Copy edit to improve clarity wrt AWS access
Previously, the title was a little too broad, which could be
misunderstood to mean that customers do not have to provide any sort of
write-access whatsoever.

This edit makes the title more specific by limiting the adjective
"read-only" to the plural noun phrase "support roles".

In addition, I have edited the intro paragraph to emphasize and clarify
that we provide access transparency.

I made some minor edits to the list of key points and added a new bullet
point mentioning the audit logging provided by AWS CloudTrail.

Finally, I added an info box below the key points that explicitly
mentions the management role as succinctly as possible, inviting the
reader to continue reading for more specifics.

As a bonus, I edited the corresponding page for read-only support access
to GCP accounts to keep the introductions consistent. Additional changes
to the GCP page are out-of-scope for this change.