Remove "read-only" from support access titles
This commit removes the "read-only" wording from the titles of both
pages relating to support access.

In addition, I have copyedited the GCP support access page for style and
consistency with the AWS page.

To match the titles, I changed the page URLs to remove the `ro`
component and added redirects to the `.gitbook.yaml` file.