Add orphaned images check and archive images
This commit moves the previously commented out orphaned images check out
of the `misc.sh` script and into the `images.sh` script, as I plan to
add additional checks for the images and figure that I may frequently
want to run this check in isolation.

The new `images.sh` script has been integrated into the `Makefile` so
that you can run it with `make images` and the GitHub Actions QA
workflow will run the check.

The check identified two kinds of error:

- Orphaned images (as designed)

- Links to images that use URL percent encoding, which trips up the
  image checker

To fix these issues, I:

- Moved all orphaned images to a new `archive/assets` directory so that
  we still have them on hand in case we need to revert this work or
  refer to them later

- Modified all image links using percent encoding to use the `<>`
  syntax, allowing whitespaces in URLs, allowing the image checker to
  function

In a future commit, I will add a check for images with spaces in them.
All images files should use hyphens instead of spaces. However, so many
changes are required to bring the image situation up to scratch that I
wish to decompose my changes into logical chunks.