Add approximate image duplicate check
This commit adds a new tool (imgdup2go) to the devcontainer setup that
can be used to check for approximate image duplicates. This tool has
been integrated into the assets check and is run against the main `docs`
directory as well as the `archive` directory.

Along with these changes, I have fixed the few duplicate issues found
and implemented ignore list functionality.

Unfortunately, this tool is quite slow. It will do for now, but I would
like to find a faster alternative.