Add image compression checking
This commit adds the optipng tool to the devcontainer setup and a CI
check that uses this tool to identify whether images can be compressed.
If any images can be compressed, the CI checks will fail.

For now, it is up to the maintainer to compress images as needed. In the
future, this can be automated with a GitHub Actions workflow that
creates new PRs.

This commit also compresses every image that can be compressed. Some
compression percentages go as high as 40%.

Additionally, the CI checks themselves have been refactored, and some
common code has been moved into the new `common.sh` script.