Add namespacing for image filenames
I have a few tasks requiring me to perform an audit of the existing
screenshots so that I can update any that are affected by changes to the
new navigation.

My recent changes to clean out the `assets` directory are a big help,
but I still have to go through a few hundred images one by one to
complete a full audit.

This change improves the situation by namepacing every asset filename
with a slug that corresponds to the type of the asset:

- `cmp` — A screenshott of the CMP
- `aws` — A screenshott of the AWS console
- `gcp` - A screenshott of the GCP console
- `ms` - A screenshott of a Microsoft product
- `slack` - A screenshott of a Slack product
- `cloudhealth` - A screenshott of CloudHealth
- `feedback-hub` - A screenshott of the Feedback Hub
- `email` - A screenshott of an email
- `file` - A file download

This commit also adds some TODO markers to the CI scripts identifying
future work that is outside of the scope of this commit.